overlapping widgets initial work

+++ src/data/layout.rs
@@ -3,5 +3,5 @@ use crate::Rect;
 #[derive(Debug, Clone, Copy)]
 pub struct Layout {
 	pub rect: Rect,
-	pub depth: f64,
+	pub depth: f32,
 }
